Spaces
It appears that the primary sails repo generates code with spaces. However, sails-generate-model
appears to use tabs.  The result is this weird mix of tabs and spaces in generated code:

https://www.dropbox.com/s/p915vh6wip9qllx/Screenshot%202014-02-15%2011.59.42.png

To keep consistency, replace the hard tabs in the templates with spaces. The rest of the code,
that is the non template code, I am indifferent about.